pyproject.toml
[project]
name = "wingify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.6.2",
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "fastapi>=0.133.0",
    "openai>=1.83.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.10",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.22",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]
